feat: Long -> Integer

+++ src/main/java/edu/allinone/sugang/repository/EnrollmentRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface EnrollmentRepository extends JpaRepository<Enrollment, Long> {
+public interface EnrollmentRepository extends JpaRepository<Enrollment, Integer> {
     // 학생 아이디와 과목 아이디로 수강 신청 정보 가져오기
-    Optional<Enrollment> findByStudentIdAndLectureId(Long studentId, Long lectureId);
+    Optional<Enrollment> findByStudentIdAndLectureId(Integer studentId, Integer lectureId);
 }
